+++ ReviewBaseSystem-ejb/src/java/ejb/clientejbLocal.java
@@ -31,9 +31,9 @@ public interface clientejbLocal {
 
     Collection<Users> getUserByCity(String city);
 
-    void addUser(String name,String email,String phno,String interest,String gender,Date bdate,String city,String pass);
+    void addUser(String name, String email, String phno, String interest, String gender, Date bdate, String city, String pass);
 
-    void updateUser(int uid,String name,String email,String phno,String interest,String gender,Date bdate,String city);
+    void updateUser(int uid, String name, String email, String phno, String interest, String gender, Date bdate, String city);
 
     void removeUser(int userId);
 
@@ -45,13 +45,15 @@ public interface clientejbLocal {
     Advertise getAdvertiseById(int advertiseId);
 
     Advertise getAdvertiseByProductId(int productId);
-    Collection<Advertise> getAdvertiseByDate(Date date); 
+
+    Collection<Advertise> getAdvertiseByDate(Date date);
 
     //Advertise getAdvertiseByDate(Date date);
-
     //(if advertise is exists between start date and end date)
-    void addAdvertise(Date startDate,Date endDate,int productId);
-    void updateAdvertise(int advertiseId,Date startDate,Date endDate,int productId);
+    void addAdvertise(Date startDate, Date endDate, int productId);
+
+    void updateAdvertise(int advertiseId, Date startDate, Date endDate, int productId);
+
     void removeAdvertise(int advertiseId);
 
     //Reviews
@@ -82,10 +84,11 @@ public interface clientejbLocal {
 
     Reviewxcriteria getReviewxCriteriaByReviewId(int reviewId);
 
-    void addReviewxCriteria(float rate, String description, int categoryRatingCriteriaId, int reviewId);
+    void addReviewxCriteria(int rate, String description, int categoryRatingCriteriaId, int reviewId);
 
-    void updateReviewxCriteria(int reviewXcriteriaId, float rate, String description, int categoryRatingCriteriaId, int reviewId);
+    void updateReviewxCriteria(int reviewXcriteriaId, int rate, String description, int categoryRatingCriteriaId, int reviewId);
 
     void removeReviewxCriteria(int reviewXcriteriaId, int categoryRatingCriteriaId, int reviewId);
+
     public Collection<Advertise> getAdvertiseByDate(String tdate);
 }
